+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Deprecate the Db class (#234)
 - Deprecate Db::enableQueryLogging (#231)
 - Mark Db::getDatabaseConnection for removal in version 4.0, not 3.0 (#228)
+- Deprecate some methods in the DataMapper (#251)
 
 ### Removed
 - Delete unused PNG file from the tests (#247)
@@ -30,7 +31,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Move `Tests/` to the dev autoload in `ext_emconf.php` (#249)
 - Keep development files out of the packages (#248)
 - Add a dependency to cms/lang in composer.json (#243)
-- Use the ConnectionPool for DB queries in TYPO3 >= 8.4 (#233, #239, #240, #242)
+- Use the ConnectionPool for DB queries in TYPO3 >= 8.4 (#233, #239, #240, #242, #251)
 - Provide flags to htmlspecialchars (#232)
 - Avoid deprecated TimeTracker usage in the testing framework (#229)
 - Clean up the testing extensions (#224)